Added tests and fixed small errors
